+++ Source/HMDissection/HMDissection.csproj
@@ -9,11 +9,12 @@
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>HMDissection</RootNamespace>
     <AssemblyName>HMDissection</AssemblyName>
-    <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v4.8.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
     <NuGetPackageImportStamp>
     </NuGetPackageImportStamp>
+    <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>false</DebugSymbols>
@@ -35,45 +36,24 @@
     <Prefer32Bit>false</Prefer32Bit>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="0Harmony">
-      <HintPath>..\References\0Harmony.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="Assembly-CSharp">
-      <HintPath>..\..\..\..\..\Program Files (x86)\Steam\steamapps\common\RimWorld\RimWorldWin64_Data\Managed\Assembly-CSharp.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="HugsLib, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\References\HugsLib.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Runtime.InteropServices.RuntimeInformation" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Xml" />
-    <Reference Include="UnityEngine.CoreModule">
-      <HintPath>..\..\..\..\..\Program Files (x86)\Steam\steamapps\common\RimWorld\RimWorldWin64_Data\Managed\UnityEngine.CoreModule.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
-  </ItemGroup>
-  <ItemGroup>
     <Compile Include="CompatibilityUtility.cs" />
     <Compile Include="DissectionDefOf.cs" />
     <Compile Include="DissectionUtility.cs" />
-    <Compile Include="CustomWidget.cs" />
-    <Compile Include="Listing_Dissection.cs" />
+    <Compile Include="Harmony\OptionalPatches\PostMortem_WorkGiver_DoBill_JobOnThing_Patch.cs" />
+    <Compile Include="Utilities\CustomWidget.cs" />
+    <Compile Include="Harmony\Pawn_WorkSettings_EnableAndInitialize_Patch.cs" />
+    <Compile Include="Harmony\Pawn_HealthTracker_Notify_Resurrected_Patch.cs" />
+    <Compile Include="Harmony\GameInitData_PrepForMapGen_Patch.cs" />
+    <Compile Include="Harmony\WorkGiver_DoBill_TryStartNewDoBillJob_Patch.cs" />
+    <Compile Include="Utilities\Listing_Dissection.cs" />
     <Compile Include="ModSettings_Dissection.cs" />
     <Compile Include="Mod_Dissection.cs" />
     <Compile Include="Filters\DissectedFilter.cs" />
-    <Compile Include="HarmonyPatches.cs" />
     <Compile Include="JobDriver_DoDissectionBill.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Toils_Dissection.cs" />
     <Compile Include="WorkGiver_DoDissectionBill.cs" />
-    <Compile Include="WorkSettingsPatches.cs" />
+    <Compile Include="Utilities\WorkSettingsPatches.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Krafs.Publicizer">
@@ -84,7 +64,11 @@
     <PackageReference Include="Krafs.Rimworld.Ref">
       <Version>1.5.4146-beta</Version>
     </PackageReference>
+    <PackageReference Include="Lib.Harmony">
+      <Version>2.3.3</Version>
+    </PackageReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>
