+++ Source/HMDissection/Harmony/Pawn_WorkSettings_EnableAndInitialize_Patch.cs
@@ -1,0 +1,16 @@
+﻿using HarmonyLib;
+using RimWorld;
+using Verse;
+
+namespace HMDissection
+{
+    [HarmonyPatch(typeof(Pawn_WorkSettings), "EnableAndInitialize")]
+    public static class Pawn_WorkSettings_EnableAndInitialize_Patch
+    {
+        public static void Postfix(Pawn ___pawn)
+        {
+            WorkSettingsPatches.SetDissectionPriorityForPassionatedPawns(___pawn);
+        }
+        
+    }
+}
