+++ Source/HMDissection/Utilities/CustomWidget.cs
@@ -1,15 +1,11 @@
 ﻿using RimWorld;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnityEngine;
 using Verse;
 using Verse.Sound;
 
 namespace HMDissection
 {
+    [StaticConstructorOnStartup]
     public static class CustomWidget
     {
